refactor(packages/contracts): organize playback request/response contracts
