+++ packages/contracts/src/playback/index.ts
@@ -1,34 +1,7 @@
-import { z } from "zod";
+// Requests
+export * from "./request/get-playback-state.request";
+export * from "./request/set-playback-state.reqeust";
 
-export const PlaybackStateSchema = z.object({
-  sessionId: z.string().nonempty(),
-  userId: z.string().nonempty(),
-  deviceName: z.string(),
-  deviceIcon: z.enum([
-    "desktop",
-    "mobile",
-    "tablet",
-    "speaker",
-    "tv",
-    "game-console",
-    "other",
-  ]),
-  isPlaying: z.boolean(),
-  trackData: z.object({
-    id: z.string(),
-    title: z.string(),
-    artists: z.array(z.string()),
-    albumArt: z.string(),
-    albumName: z.string(),
-    albumId: z.string(),
-    duration: z.number().min(0).int(),
-  }),
-  currentTime: z.number().min(0).int(),
-  volume: z.number().min(0).max(1),
-  repeatMode: z.enum(["off", "all", "one"]),
-  shuffle: z.boolean(),
-  favorited: z.enum(["favorited", "disliked", "not-set"]),
-  inLibrary: z.boolean(),
-});
-
-export type PlaybackState = z.infer<typeof PlaybackStateSchema>;
+// Responses
+export * from "./response/get-playback-state.response";
+export * from "./response/set-playback-state.response";
