+++ packages/contracts/src/playback/request/get-playback-state.request.ts
@@ -1,0 +1,7 @@
+import { z } from "zod";
+
+export const GetPlaybackStateRequestSchema = z.object({}).strict();
+
+export type GetPlaybackStateRequest = z.infer<
+  typeof GetPlaybackStateRequestSchema
+>;
