+++ packages/contracts/src/playback/response/get-playback-state.response.ts
@@ -1,0 +1,8 @@
+import { z } from "zod";
+import { PlaybackStateSchema } from "../../schemas/playback.schema";
+
+export const GetPlaybackStateResponseSchema = PlaybackStateSchema;
+
+export type GetPlaybackStateResponse = z.infer<
+  typeof GetPlaybackStateResponseSchema
+>;
