+++ packages/contracts/src/playback/response/set-playback-state.response.ts
@@ -1,0 +1,8 @@
+import { z } from "zod";
+import { PlaybackStateSchema } from "../../schemas/playback.schema";
+
+export const SetPlaybackStateResponseSchema = PlaybackStateSchema;
+
+export type SetPlaybackStateResponse = z.infer<
+  typeof SetPlaybackStateResponseSchema
+>;
