+++ packages/contracts/src/schemas/index.ts
@@ -23,6 +23,7 @@ export * from "./moderator-account.schema";
 export * from "./notification.schema";
 export * from "./playlist-track.schema";
 export * from "./playlist.schema";
+export * from "./playback.schema";
 export * from "./refresh-token.schema";
 export * from "./report-target.schema";
 export * from "./report.schema";
